Restructure and clean up of main logic, remove hooks
